Feat: add Translator parent class
