+++ lib/to_latin.rb
@@ -1,13 +1,5 @@
-require 'pry'
-class ToLatin
-  attr_reader :dictionary
-  attr_accessor :text
-
-  def initialize(text, dictionary)
-    @text = text
-    @dictionary = dictionary
-  end
-
+require './lib/translator'
+class ToLatin < Translator
   def collect_rows
     @text.split("\n")
   end
